+++ Sources/SuperwallKit/Config/Models/Config.swift
@@ -12,7 +12,6 @@ struct Config: Decodable {
   var triggers: Set<Trigger>
   var paywalls: [Paywall]
   var logLevel: Int
-  var postback: PostbackRequest
   var locales: Set<String>
   var appSessionTimeout: Milliseconds
   var featureFlags: FeatureFlags
@@ -31,7 +30,6 @@ struct Config: Decodable {
     case triggers = "triggerOptions"
     case paywalls = "paywallResponses"
     case logLevel
-    case postback
     case localization
     case appSessionTimeout = "appSessionTimeoutMs"
     case featureFlags = "toggles"
@@ -45,7 +43,6 @@ struct Config: Decodable {
     triggers = try values.decode(Set<Trigger>.self, forKey: .triggers)
     paywalls = try values.decode([Paywall].self, forKey: .paywalls)
     logLevel = try values.decode(Int.self, forKey: .logLevel)
-    postback = try values.decode(PostbackRequest.self, forKey: .postback)
     appSessionTimeout = try values.decode(Milliseconds.self, forKey: .appSessionTimeout)
     featureFlags = try FeatureFlags(from: decoder)
     preloadingDisabled = try values.decode(PreloadingDisabled.self, forKey: .preloadingDisabled)
@@ -59,7 +56,6 @@ struct Config: Decodable {
     triggers: Set<Trigger>,
     paywalls: [Paywall],
     logLevel: Int,
-    postback: PostbackRequest,
     locales: Set<String>,
     appSessionTimeout: Milliseconds,
     featureFlags: FeatureFlags,
@@ -69,7 +65,6 @@ struct Config: Decodable {
     self.triggers = triggers
     self.paywalls = paywalls
     self.logLevel = logLevel
-    self.postback = postback
     self.locales = locales
     self.appSessionTimeout = appSessionTimeout
     self.featureFlags = featureFlags
@@ -85,7 +80,6 @@ extension Config: Stubbable {
       triggers: [.stub()],
       paywalls: [.stub()],
       logLevel: 0,
-      postback: .stub(),
       locales: [],
       appSessionTimeout: 3600000,
       featureFlags: .stub(),