+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 The changelog for `SuperwallKit`. Also see the [releases](https://github.com/superwall/Superwall-iOS/releases) on GitHub.
 
-## 3.7.5
+## 3.8.0
 
 ### Enhancements